requirements.txt
streamlit
pandas
scikit-learn
joblib
nltk
matplotlib
pdfplumber
pillow
pytesseract
scipy
